Moved contract functions to utils folder
